Add files via upload
Code to get the metadata of the songs I played during the last year on my Spotify profile from Spotify API. This code was originally created here: https://github.com/vlad-ds/spoty-records. I decided to grab his code and put it in a notebook to have a better understanding of what the code was doing step by step.